+++ sider/hash.py
@@ -11,6 +11,7 @@
 import collections
 from .session import Session
 from .types import Bulk, ByteString
+from .transaction import query, manipulative
 
 
 class Hash(collections.MutableMapping):
@@ -62,6 +63,7 @@ class Hash(collections.MutableMapping):
         self.value_type = Bulk.ensure_value_type(value_type,
                                                  parameter='value_type')
 
+    @query
     def __iter__(self):
         """Iterates over its :meth:`keys()`.
 
@@ -78,6 +80,7 @@ class Hash(collections.MutableMapping):
         for key in keys:
             yield decode(key)
 
+    @query
     def __len__(self):
         """Gets the number of items.
 
@@ -91,6 +94,7 @@ class Hash(collections.MutableMapping):
         """
         return self.session.client.hlen(self.key)
 
+    @query
     def __contains__(self, key):
         """Tests whether the given ``key`` exists.
 
@@ -110,6 +114,7 @@ class Hash(collections.MutableMapping):
         exists = self.session.client.hexists(self.key, encoded_key)
         return bool(exists)
 
+    @query
     def __getitem__(self, key):
         """Gets the value of the given ``key``.
 
@@ -132,6 +137,7 @@ class Hash(collections.MutableMapping):
             raise KeyError(key)
         return self.value_type.decode(value)
 
+    @manipulative
     def __setitem__(self, key, value):
         """Sets the ``key`` with the ``value``.
 
@@ -166,8 +172,16 @@ class Hash(collections.MutableMapping):
            It is directly mapped to Redis :redis:`HDEL` command.
 
         """
+        session = self.session
         encoded = self.key_type.encode(key)
-        ok = self.session.client.hdel(self.key, encoded)
+        if session.current_transaction is None:
+            ok = session.client.hdel(self.key, encoded)
+        else:
+            session.mark_query([self.key])
+            ok = session.client.hexists(self.key, encoded)
+            if ok:
+                session.mark_manipulative()
+                session.client.hdel(self.key, encoded)
         if not ok:
             raise KeyError(key)
 
@@ -187,6 +201,7 @@ class Hash(collections.MutableMapping):
         """
         return frozenset(self)
 
+    @query
     def values(self):
         """Gets its all values.  It returns a :class:`list` but
         there isn't any meaningful order of values.
@@ -206,6 +221,7 @@ class Hash(collections.MutableMapping):
             values[i] = decode(val)
         return values
 
+    @query
     def items(self):
         """Gets its all ``(key, value)`` pairs.
         There isn't any meaningful order of pairs.
@@ -224,6 +240,7 @@ class Hash(collections.MutableMapping):
         return frozenset((decode_key(k), decode_value(v))
                          for k, v in items.iteritems())
 
+    @manipulative
     def clear(self):
         """Removes all items from this hash.
 
@@ -266,6 +283,14 @@ class Hash(collections.MutableMapping):
            command which is atomic.
 
         """
+        if self.session.transaction is not None:
+            self.session.mark_query()
+            try:
+                val = self[key]
+            except KeyError:
+                self.session.mark_manipulative([self.key])
+                self[key] = val = default
+            return val
         encoded_key = self.key_type.encode(key)
         encoded_val = self.value_type.encode(default)
         result = [None]
@@ -280,6 +305,7 @@ class Hash(collections.MutableMapping):
         self.session.client.transaction(block, self.key)
         return result[0]
 
+    @manipulative
     def update(self, mapping={}, **keywords):
         """Updates the hash from the given ``mapping`` and keyword
         arguments.
@@ -337,4 +363,11 @@ class Hash(collections.MutableMapping):
                            for val in (encode_key(k), encode_value(v)))
         pipe.execute_command('HMSET', self.key, *flatten)
 
+    def __repr__(self):
+        cls = type(self)
+        items = list(self.iteritems())
+        items.sort(key=lambda (key, _): key)
+        elements = ', '.join('{0!r}: {1!r}'.format(*pair) for pair in items)
+        return '<{0}.{1} ({2!r}) {{{3}}}>'.format(cls.__module__, cls.__name__,
+                                                  self.key, elements)